Added GraphQl support

+++ src/main/java/com/pluralsight/service/ApplicationService.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface ApplicationService {
     List<Application> listApplications();
-    Application findApplication(Integer id);
+    Application findApplication(Long id);
 }
